backend-functions-docker/requirements.txt
flask
playwright
beautifulsoup4
openai
python-dotenv
requests
azure-functions 